Fix up imports + remove hardcoded tiles url
