+++ shared/studio/package.json
@@ -20,9 +20,12 @@
     "fuzzysort": "^2.0.1",
     "graphql": "^16.5.0",
     "idb": "^7.0.1",
+    "maplibre-gl": "^4.5.2",
     "mobx": "^6.5.0",
     "mobx-keystone": "^1.11.0",
     "mobx-react-lite": "^4.0.0",
+    "pmtiles": "^3.0.7",
+    "protomaps-themes-base": "^3.1.0",
     "react-colorful": "^5.6.1",
     "react-error-boundary": "^3.1.4",
     "react-hook-form": "^7.32.2",
