+++ web/package.json
@@ -44,9 +44,6 @@
     "generic-names": "^4.0.0",
     "globals": "^15.11.0",
     "jest": "^29.7.0",
-    "maplibre-gl": "^4.5.2",
-    "pmtiles": "^3.0.7",
-    "protomaps-themes-base": "^3.1.0",
     "sass": "^1.49.9",
     "selenium-webdriver": "^4.26.0",
     "ts-jest": "^29.2.5",
